+++ packages/descriptions/src/useFetchData.tsx
@@ -55,11 +55,8 @@ const useFetchData = <T extends RequestData>(
       }
     } catch (e) {
       // 如果没有传递这个方法的话，需要把错误抛出去，以免吞掉错误
-      if (onRequestError === undefined) {
-        throw new Error(e as string);
-      } else {
-        onRequestError(e as Error);
-      }
+      if (onRequestError === undefined) throw new Error(e as string);
+      else onRequestError(e as Error);
       setLoading(false);
     }
   };