+++ packages/form/src/components/StepsForm/demos/modal-step-form.tsx
@@ -41,7 +41,6 @@ export default () => {
         }}
         stepsFormRender={(dom, submitter) => {
           return (
-            // @ts-expect-error
             <Modal
               title="分步表单"
               width={800}