+++ packages/form/src/layouts/DrawerForm/index.tsx
@@ -9,6 +9,7 @@ import type { CommonFormProps } from '../../BaseForm';
 import { BaseForm } from '../../BaseForm';
 import { noteOnce } from 'rc-util/lib/warning';
 import merge from 'lodash/merge';
+import { useRefFunction } from '@ant-design/pro-utils';
 
 export type DrawerFormProps<T = Record<string, any>> = Omit<FormProps, 'onFinish' | 'title'> &
   CommonFormProps<T> & {
@@ -114,6 +115,7 @@ function DrawerForm<T = Record<string, any>>({
     if (rest.submitter === false) {
       return false;
     }
+
     return merge(
       {
         searchConfig: {
@@ -151,27 +153,22 @@ function DrawerForm<T = Record<string, any>>({
     );
   }, []);
 
-  const onFinishHandle = useCallback(
-    async (values: T) => {
-      const response = onFinish?.(values);
-
-      if (submitTimeout && response instanceof Promise) {
-        setLoading(true);
-
-        const timer = setTimeout(() => setLoading(false), submitTimeout);
-        response.finally(() => {
-          clearTimeout(timer);
-          setLoading(false);
-        });
-      }
-      const result = await response;
-      // 返回真值，关闭弹框
-      if (result) {
-        setVisible(false);
-      }
-    },
-    [onFinish, setVisible, submitTimeout],
-  );
+  const onFinishHandle = useRefFunction(async (values: T) => {
+    const response = onFinish?.(values);
+    if (submitTimeout && response instanceof Promise) {
+      setLoading(true);
+      const timer = setTimeout(() => setLoading(false), submitTimeout);
+      response.finally(() => {
+        clearTimeout(timer);
+        setLoading(false);
+      });
+    }
+    const result = await response;
+    // 返回真值，关闭弹框
+    if (result) {
+      setVisible(false);
+    }
+  });
 
   return (
     <>