+++ tests/form/lightFilter.test.tsx
@@ -1,4 +1,4 @@
-import { mount, render } from 'enzyme';
+import { mount } from 'enzyme';
 import React from 'react';
 import {
   LightFilter,
@@ -14,7 +14,9 @@ import {
 import KeyCode from 'rc-util/lib/KeyCode';
 import { act } from 'react-dom/test-utils';
 import { waitForComponentToPaint } from '../util';
+import { render } from '@testing-library/react';
 import moment from 'moment';
+import '@testing-library/jest-dom';
 
 describe('LightFilter', () => {
   it(' 🪕 basic use', async () => {
@@ -667,7 +669,7 @@ describe('LightFilter', () => {
     });
   });
 
-  it('ProFormField support lightProps', () => {
+  it('🪕 ProFormField support lightProps', () => {
     const html = render(
       <LightFilter
         initialValues={{
@@ -699,12 +701,11 @@ describe('LightFilter', () => {
       </LightFilter>,
     );
 
-    expect(html.find('.ant-pro-core-field-label').text()).toBe(
-      '活跃时间: 2001-09-09 01:46:40~2017-07-14 0...2项',
-    );
+    const inputDom = html.findAllByText('活跃时间: 2001-09-09 01:46:40~2017-07-14 0...2项');
+    expect(!!inputDom).toBeTruthy();
   });
 
-  it('lightFilter support placement', async () => {
+  it('🪕 lightFilter support placement', async () => {
     const wrapper = mount(
       <LightFilter
         initialValues={{