+++ tests/form/modalForm.test.tsx
@@ -1,15 +1,14 @@
 ﻿import React from 'react';
 import { ProFormText, ModalForm } from '@ant-design/pro-form';
 import { Button } from 'antd';
-import { act } from 'react-dom/test-utils';
-import { render } from '@testing-library/react';
-import { mount } from 'enzyme';
+import { fireEvent, act, render } from '@testing-library/react';
+import '@testing-library/jest-dom';
 import { waitForComponentToPaint, waitTime } from '../util';
 
 describe('ModalForm', () => {
   it('📦 trigger will simulate onVisibleChange', async () => {
     const fn = jest.fn();
-    const wrapper = mount(
+    const wrapper = render(
       <ModalForm
         width={600}
         trigger={<Button id="new">新建</Button>}
@@ -20,8 +19,8 @@ describe('ModalForm', () => {
     );
     await waitForComponentToPaint(wrapper);
 
-    act(() => {
-      wrapper.find('button#new').simulate('click');
+    await act(async () => {
+      (await wrapper.findByText('新 建'))?.click();
     });
     await waitForComponentToPaint(wrapper);
     expect(fn).toBeCalledWith(true);
@@ -29,7 +28,7 @@ describe('ModalForm', () => {
 
   it('📦 submitter config no reset default config', async () => {
     const fn = jest.fn();
-    const wrapper = mount(
+    const wrapper = render(
       <ModalForm
         width={600}
         submitter={{
@@ -52,14 +51,14 @@ describe('ModalForm', () => {
     );
     await waitForComponentToPaint(wrapper);
 
-    act(() => {
-      wrapper.find('button#new').simulate('click');
+    await act(async () => {
+      (await wrapper.findByText('新 建'))?.click();
     });
     await waitForComponentToPaint(wrapper, 200);
     expect(fn).toBeCalledWith(true);
 
-    act(() => {
-      wrapper.find('button#reset').simulate('click');
+    await act(async () => {
+      (await wrapper.findByText('取 消'))?.click();
     });
     await waitForComponentToPaint(wrapper);
     expect(fn).toBeCalledWith(false);
@@ -67,38 +66,12 @@ describe('ModalForm', () => {
 
   it('📦 ModalForm first no render items', async () => {
     const fn = jest.fn();
-    const wrapper = mount(
+    const wrapper = render(
       <ModalForm
         width={600}
         trigger={<Button id="new">新建</Button>}
-        onVisibleChange={(visible) => fn(visible)}
-      >
-        <ProFormText
-          name="name"
-          fieldProps={{
-            id: 'test',
-          }}
-        />
-      </ModalForm>,
-    );
-    await waitForComponentToPaint(wrapper);
-
-    expect(wrapper.find('input#test').exists()).toBeFalsy();
-
-    act(() => {
-      wrapper.find('button#new').simulate('click');
-    });
-    await waitForComponentToPaint(wrapper);
-    expect(wrapper.find('input#test').exists()).toBeTruthy();
-  });
-
-  it('📦 ModalForm first render items', async () => {
-    const fn = jest.fn();
-    const wrapper = mount(
-      <ModalForm
-        width={600}
-        modalProps={{
-          forceRender: true,
+        initialValues={{
+          name: '1234',
         }}
         onVisibleChange={(visible) => fn(visible)}
       >
@@ -112,14 +85,50 @@ describe('ModalForm', () => {
     );
     await waitForComponentToPaint(wrapper);
 
-    expect(wrapper.find('input#test').exists()).toBeTruthy();
+    expect(await wrapper.queryByDisplayValue('1234')).toBeFalsy();
+
+    await act(async () => {
+      (await wrapper.findByText('新 建'))?.click();
+    });
+    await waitForComponentToPaint(wrapper);
+    expect(await wrapper.findByDisplayValue('1234')).toBeTruthy();
+  });
+
+  it('📦 ModalForm first render items', async () => {
+    const fn = jest.fn();
+    const wrapper = render(
+      <ModalForm
+        width={600}
+        modalProps={{
+          forceRender: true,
+        }}
+        initialValues={{
+          name: '1234',
+        }}
+        visible
+        onVisibleChange={(visible) => fn(visible)}
+      >
+        <ProFormText
+          name="name"
+          fieldProps={{
+            id: 'test',
+          }}
+        />
+      </ModalForm>,
+    );
+    await waitForComponentToPaint(wrapper, 120);
+
+    expect(await wrapper.findByDisplayValue('1234')).toBeTruthy();
   });
 
   it('📦 ModalForm destroyOnClose', async () => {
     const fn = jest.fn();
-    const wrapper = mount(
+    const wrapper = render(
       <ModalForm
         width={600}
+        initialValues={{
+          name: '1234',
+        }}
         modalProps={{ destroyOnClose: true }}
         onVisibleChange={(visible) => fn(visible)}
       >
@@ -133,30 +142,61 @@ describe('ModalForm', () => {
     );
     await waitForComponentToPaint(wrapper);
 
-    expect(wrapper.find('input#test').exists()).toBeFalsy();
+    expect(await wrapper.queryByDisplayValue('1234')).toBeFalsy();
 
     act(() => {
-      wrapper.setProps({
-        visible: true,
-      });
+      wrapper.rerender(
+        <ModalForm
+          width={600}
+          initialValues={{
+            name: '1234',
+          }}
+          visible
+          modalProps={{ destroyOnClose: true }}
+          onVisibleChange={(visible) => fn(visible)}
+        >
+          <ProFormText
+            name="name"
+            fieldProps={{
+              id: 'test',
+            }}
+          />
+        </ModalForm>,
+      );
     });
     await waitForComponentToPaint(wrapper);
 
-    expect(wrapper.find('input#test').exists()).toBeTruthy();
+    expect(await wrapper.findByDisplayValue('1234')).toBeTruthy();
 
     act(() => {
-      wrapper.setProps({
-        visible: false,
-      });
+      wrapper.rerender(
+        <ModalForm
+          key="reset"
+          width={600}
+          initialValues={{
+            name: '1234',
+          }}
+          visible={false}
+          modalProps={{ destroyOnClose: true }}
+          onVisibleChange={(visible) => fn(visible)}
+        >
+          <ProFormText
+            name="name"
+            fieldProps={{
+              id: 'test',
+            }}
+          />
+        </ModalForm>,
+      );
     });
     await waitForComponentToPaint(wrapper, 2000);
 
-    expect(wrapper.find('input#test').exists()).toBeFalsy();
+    expect(await wrapper.queryByDisplayValue('1234')).toBeFalsy();
   });
 
   it('📦 modal close button will simulate onVisibleChange', async () => {
     const fn = jest.fn();
-    const wrapper = mount(
+    const wrapper = render(
       <ModalForm
         visible
         trigger={<Button id="new">新建</Button>}
@@ -168,7 +208,7 @@ describe('ModalForm', () => {
     await waitForComponentToPaint(wrapper);
 
     act(() => {
-      wrapper.find('button.ant-modal-close').simulate('click');
+      wrapper.baseElement.querySelector<HTMLDivElement>('button.ant-modal-close')?.click();
     });
     await waitForComponentToPaint(wrapper);
     expect(fn).toBeCalledWith(false);
@@ -177,7 +217,7 @@ describe('ModalForm', () => {
 
   it('📦 modal visible=true simulate onVisibleChange', async () => {
     const fn = jest.fn();
-    const wrapper = mount(
+    const wrapper = render(
       <ModalForm
         visible
         trigger={<Button id="new">新建</Button>}
@@ -193,7 +233,7 @@ describe('ModalForm', () => {
 
   it('📦 reset button will simulate onVisibleChange', async () => {
     const fn = jest.fn();
-    const wrapper = mount(
+    const wrapper = render(
       <ModalForm
         visible
         trigger={<Button id="new">新建</Button>}
@@ -204,15 +244,15 @@ describe('ModalForm', () => {
     );
     await waitForComponentToPaint(wrapper);
 
-    act(() => {
-      wrapper.find('.ant-modal-footer').update().find('button.ant-btn').at(0).simulate('click');
+    await act(async () => {
+      (await wrapper.findByText('取 消'))?.click();
     });
     expect(fn).toBeCalledWith(false);
   });
 
   it('📦 modal close button will simulate modalProps.onCancel', async () => {
     const fn = jest.fn();
-    const wrapper = mount(
+    const wrapper = render(
       <ModalForm
         visible
         modalProps={{
@@ -227,7 +267,7 @@ describe('ModalForm', () => {
     await waitForComponentToPaint(wrapper);
 
     act(() => {
-      wrapper.find('button.ant-modal-close').simulate('click');
+      wrapper.baseElement.querySelector<HTMLDivElement>('button.ant-modal-close')?.click();
     });
     await waitForComponentToPaint(wrapper);
     expect(fn).toBeCalledWith(false);
@@ -235,7 +275,7 @@ describe('ModalForm', () => {
 
   it('📦 form onFinish return true should close modal', async () => {
     const fn = jest.fn();
-    const wrapper = mount(
+    const wrapper = render(
       <ModalForm
         visible
         trigger={<Button id="new">新建</Button>}
@@ -247,8 +287,8 @@ describe('ModalForm', () => {
     );
     await waitForComponentToPaint(wrapper, 500);
 
-    act(() => {
-      wrapper.find('button.ant-btn-primary').simulate('click');
+    await act(async () => {
+      (await wrapper.findByText('确 认'))?.click();
     });
 
     await waitForComponentToPaint(wrapper);
@@ -258,7 +298,7 @@ describe('ModalForm', () => {
 
   it('📦 form onFinish is null, no close modal', async () => {
     const fn = jest.fn();
-    const wrapper = mount(
+    const wrapper = render(
       <ModalForm
         visible
         trigger={<Button id="new">新建</Button>}
@@ -269,8 +309,8 @@ describe('ModalForm', () => {
     );
     await waitForComponentToPaint(wrapper, 500);
 
-    act(() => {
-      wrapper.find('button.ant-btn-primary').simulate('click');
+    await act(async () => {
+      (await wrapper.findByText('确 认'))?.click();
     });
 
     await waitForComponentToPaint(wrapper);
@@ -278,24 +318,23 @@ describe('ModalForm', () => {
   });
 
   it('📦 ModalForm support submitter is false', async () => {
-    const wrapper = mount(
+    const wrapper = render(
       <ModalForm visible trigger={<Button id="new">新建</Button>} submitter={false}>
         <ProFormText name="name" />
       </ModalForm>,
     );
     await waitForComponentToPaint(wrapper);
 
-    act(() => {
-      wrapper.find('button#new').simulate('click');
+    await act(async () => {
+      (await wrapper.findByText('新 建'))?.click();
     });
 
     await waitForComponentToPaint(wrapper);
-
-    expect(wrapper.find('.ant-modal-footer').length).toBe(0);
+    expect(wrapper.baseElement.querySelector<HTMLDivElement>('.ant-modal-footer')).toBeFalsy();
   });
 
   it('📦 ModalForm close no rerender from', async () => {
-    const wrapper = mount(
+    const wrapper = render(
       <ModalForm
         initialValues={{
           name: '1234',
@@ -312,39 +351,37 @@ describe('ModalForm', () => {
     );
     await waitForComponentToPaint(wrapper);
 
-    act(() => {
-      wrapper.find('button#new').simulate('click');
+    await act(async () => {
+      (await wrapper.findByText('新 建'))?.click();
     });
 
     await waitForComponentToPaint(wrapper, 300);
+
     act(() => {
-      wrapper
-        .find('.ant-input#test')
-        .at(0)
-        .simulate('change', {
-          target: {
-            value: 'test',
-          },
-        });
+      fireEvent.change(wrapper.baseElement.querySelector('.ant-input#test')!, {
+        target: {
+          value: 'test',
+        },
+      });
     });
     await waitForComponentToPaint(wrapper);
-    expect(wrapper.find('input#test').props().value).toEqual('test');
+    expect(await wrapper.findByDisplayValue('test')).toBeTruthy();
     await waitForComponentToPaint(wrapper);
 
     act(() => {
-      wrapper.find('.ant-modal-close').simulate('click');
+      wrapper.baseElement.querySelector<HTMLDivElement>('button.ant-modal-close')?.click();
     });
     await waitForComponentToPaint(wrapper);
-    act(() => {
-      wrapper.find('button#new').simulate('click');
+    await act(async () => {
+      (await wrapper.findByText('新 建'))?.click();
     });
     await waitForComponentToPaint(wrapper);
 
-    expect(wrapper.find('input#test').props().value).toEqual('test');
+    expect(await wrapper.findByDisplayValue('test')).toBeTruthy();
   });
 
   it('📦 ModalForm destroyOnClose close will rerender from', async () => {
-    const wrapper = mount(
+    const wrapper = render(
       <ModalForm
         modalProps={{
           getContainer: false,
@@ -364,91 +401,91 @@ describe('ModalForm', () => {
       </ModalForm>,
     );
     await waitForComponentToPaint(wrapper);
-    act(() => {
-      wrapper.find('button#new').simulate('click');
+    await act(async () => {
+      (await wrapper.findByText('新 建'))?.click();
     });
 
     await waitForComponentToPaint(wrapper, 300);
     act(() => {
-      wrapper
-        .find('.ant-input#test')
-        .at(0)
-        .simulate('change', {
-          target: {
-            value: '1111',
-          },
-        });
+      fireEvent.change(wrapper.container.querySelector('.ant-input#test')!, {
+        target: {
+          value: '1111',
+        },
+      });
     });
 
     await waitForComponentToPaint(wrapper);
-    expect(wrapper.find('input#test').props().value).toEqual('1111');
+    expect(await wrapper.findByDisplayValue('1111')).toBeTruthy();
 
     await waitForComponentToPaint(wrapper);
 
     act(() => {
-      wrapper.find('.ant-modal-close').simulate('click');
+      wrapper.baseElement.querySelector<HTMLDivElement>('button.ant-modal-close')?.click();
     });
+
     await waitForComponentToPaint(wrapper);
-    act(() => {
-      wrapper.find('button#new').simulate('click');
+    await act(async () => {
+      (await wrapper.findByText('新 建'))?.click();
     });
     await waitForComponentToPaint(wrapper);
 
-    expect(wrapper.find('input#test').props().value).toEqual('1234');
+    expect(await wrapper.findByDisplayValue('1234')).toBeTruthy();
   });
 
-  it('📦 modal submitTimeout is number will disabled close button when submit', async () => {
+  it('📦 DrawerForm submitTimeout is number will disabled close button when submit', async () => {
     const fn = jest.fn();
-    const wrapper = mount(
+    const html = render(
       <ModalForm
         visible
         modalProps={{
           onCancel: () => fn(),
         }}
         onFinish={async () => {
-          await waitTime(2000);
+          await waitTime(3000);
         }}
         submitTimeout={3000}
-      />,
+      >
+        <ProFormText name="text" />
+      </ModalForm>,
     );
-    await waitForComponentToPaint(wrapper, 500);
+    await waitForComponentToPaint(html, 500);
 
-    act(() => {
-      wrapper.find('button.ant-btn-primary').simulate('click');
+    await act(async () => {
+      (await html.findByText('确 认'))?.click();
     });
 
-    await waitForComponentToPaint(wrapper, 500);
+    await waitForComponentToPaint(html, 1000);
 
-    expect(wrapper.find('button.ant-btn-default').props().disabled).toEqual(true);
+    expect(
+      (html.queryAllByText('取 消').at(0)?.parentElement as HTMLButtonElement).disabled,
+    ).toEqual(true);
 
-    act(() => {
-      wrapper.find('.ant-modal-close').simulate('click');
+    await act(async () => {
+      (await html.queryByText('取 消'))?.click();
     });
 
-    await waitForComponentToPaint(wrapper, 500);
+    await waitForComponentToPaint(html, 500);
 
     expect(fn).not.toBeCalled();
 
-    await waitForComponentToPaint(wrapper, 2500);
+    await waitForComponentToPaint(html, 2500);
 
-    expect(wrapper.find('button.ant-btn-default').props().disabled).toEqual(false);
+    expect(
+      (html.queryAllByText('取 消').at(0)?.parentElement as HTMLButtonElement)?.disabled,
+    ).toEqual(false);
 
-    act(() => {
-      wrapper.find('.ant-modal-close').simulate('click');
+    await act(async () => {
+      (await html.queryByText('取 消'))?.click();
     });
 
-    await waitForComponentToPaint(wrapper, 500);
+    await waitForComponentToPaint(html, 1000);
 
     expect(fn).toBeCalled();
-
-    act(() => {
-      wrapper.unmount();
-    });
   });
 
   it('📦 modal submitTimeout is null no disable close button when submit', async () => {
     const fn = jest.fn();
-    const wrapper = mount(
+    const wrapper = render(
       <ModalForm
         visible
         modalProps={{
@@ -461,16 +498,18 @@ describe('ModalForm', () => {
     );
     await waitForComponentToPaint(wrapper, 500);
 
-    act(() => {
-      wrapper.find('button.ant-btn-primary').simulate('click');
+    await act(async () => {
+      (await wrapper.findByText('确 认'))?.click();
     });
 
     await waitForComponentToPaint(wrapper, 500);
 
-    expect(wrapper.find('button.ant-btn-default').props().disabled).toEqual(undefined);
+    expect(
+      wrapper.baseElement.querySelector<HTMLButtonElement>('button.ant-btn-default')?.disabled,
+    ).toEqual(false);
 
     act(() => {
-      wrapper.find('.ant-modal-close').simulate('click');
+      wrapper.baseElement.querySelector<HTMLDivElement>('button.ant-modal-close')?.click();
     });
 
     await waitForComponentToPaint(wrapper, 500);
@@ -503,7 +542,7 @@ describe('ModalForm', () => {
   it('📦 ModelForm get formRef when destroyOnClose', async () => {
     const ref = React.createRef<any>();
 
-    const html = mount(
+    const html = render(
       <ModalForm
         formRef={ref}
         modalProps={{
@@ -521,8 +560,8 @@ describe('ModalForm', () => {
 
     waitForComponentToPaint(html, 200);
     expect(ref.current).toBeFalsy();
-    act(() => {
-      html.find('button#new').simulate('click');
+    await act(async () => {
+      (await html.findByText('新 建'))?.click();
     });
     await waitForComponentToPaint(html, 200);